Give Recipes an optional ID
If it's not present when the class is constructed, generate a UUIDv4 to
use as an ID. Now we have custom init logic this isn't a dataclass any
more so drop the decorator.

This ID is necessary because we'll need a unique ID to use in the
primary key for DynamoDB and there's no functionality in the database to
generate these.