Parse the results of the DynamoDB query back to Recipes
I don't love the pattern I've used to init the DynamoDB client on the
Recipe class, but I wanted to build this with a Rails style ORM interface
where the class contains the data and all the methods to search and
write to the database. I may end up refactoring later.