Copy lambda code into a new folder before zipping
The zipped archive had the contents of the `backend` folder but didn't
include a top level folder called `backend`. This caused the imports
in the lambda function to fail.

Copy the contents of the backend folder into a new folder before
deploying. This new folder is `/dist/backend/backend`. When we do the
ZIP now we package from the higher level folder which means our ZIP file
now starts with a `backend` folder.

This resource only needs to be run when there's a change to the lambda
code - updating lambdas is the slowest part of the deployment so don't
want to do that if we're only updating eg. the DynamoDB table. The
trigger on this resource calculates a hash for every file in the
`/backend` folder, concatenates them and hashes the resulting string.
See [^1] for more explanation.

[^1]: https://stackoverflow.com/questions/51138667/can-terraform-watch-a-directory-for-change