Load the plan at the top level component
Load the planned recipes in Plan and push it down to the Days. We have
one object representing all the planned recipes and we need to treat
state as immutable. This means each day needs the full copy of the plan
and will need to update its day and set the state with the updated full
plan.