Call the API client to save the plan after changes
This uses a timeout and a cleanup function in `useEffect` to call the
API client and save the updated plan 3 seconds after it's modified.

This timeout strategy is needed because `useEffect` is called every time
the state updates and on this page that includes when a user types in
the notes for a day. We don't want to DDoS our own API by making a
request on every keystroke, so this 3 second delay (after the last change)
should be a good balance between saving all the data and not breaking
the API.