Handle the response from Google Auth
The docs for the Google Auth React package are wrong - the response
isn't an access token which allows you to call the userinfo endpoint.
Instead we get a JWT directly with all the data and can simply decode
it (if we don't care about verifying the signature, which I don't here).