Handle cases when the local timezone isn't UTC
The recipes were showing as the correct day on the plan view but they
were off by a day in the database and in the Google calendar events.

I realised this is because the UK is now in UTC+1 for summertime and
the naive conversion of the Date type to a YYYY-MM-DD string puts
the time back from local to UTC. This meant every day was calculated
as the day before.